[xwq] 重构

+++ my-projrect/src/main/java/com/manage/service/SysRoleMenuService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface SysRoleMenuService {
 	
-	void saveOrUpdate(Long roleId, List<Long> menuIdList);
+	void saveOrUpdate(Integer roleId, List<Integer> menuIdList);
 	
    //根据角色ID，获取菜单ID列表
-	List<Long> queryMenuIdList(Long roleId);
+	List<Integer> queryMenuIdList(Integer roleId);
 	
 }
